refactor lsp server
