+++ Composer/packages/extensions/obiformeditor/package.json
@@ -33,8 +33,6 @@
     "classnames": "^2.2.6",
     "format-message": "^6.2.1",
     "lodash": "^4.17.15",
-    "lodash.debounce": "^4.0.8",
-    "lodash.get": "^4.4.2",
     "nanoid": "^2.0.1",
     "office-ui-fabric-react": "7.37.1",
     "react-error-boundary": "^1.2.5"
