+++ Composer/packages/server/src/server.ts
@@ -8,12 +8,16 @@ import crypto from 'crypto';
 import express, { Express, Request, Response, NextFunction } from 'express';
 import bodyParser from 'body-parser';
 import morgan from 'morgan';
-import { attachLSPServer } from '@bfc/lg-languageserver';
 import compression from 'compression';
+import * as ws from 'ws';
+import * as rpc from 'vscode-ws-jsonrpc';
+import { IConnection, createConnection } from 'vscode-languageserver';
+import { LGServer } from '@bfc/lg-languageserver';
 
 import { getAuthProvider } from './router/auth';
 import { apiRouter } from './router/api';
 import { BASEURL } from './constants';
+import { attachLSPServer } from './utility/attachLSP';
 
 const app: Express = express();
 app.set('view engine', 'ejs');
@@ -102,4 +106,28 @@ const server = app.listen(port, () => {
     console.log(`\n\nComposer now running at:\n\nhttp://localhost:${port}\n`);
   }
 });
-attachLSPServer(server, '/lgServer');
+
+// set up language server
+const wss: ws.Server = new ws.Server({
+  noServer: true,
+  perMessageDeflate: false,
+});
+
+function launchLanguageServer(socket: rpc.IWebSocket) {
+  const reader = new rpc.WebSocketMessageReader(socket);
+  const writer = new rpc.WebSocketMessageWriter(socket);
+  const connection: IConnection = createConnection(reader, writer);
+  const server = new LGServer(connection);
+  server.start();
+}
+
+attachLSPServer(wss, server, '/lgServer', webSocket => {
+  // launch language server when the web socket is opened
+  if (webSocket.readyState === webSocket.OPEN) {
+    launchLanguageServer(webSocket);
+  } else {
+    webSocket.on('open', () => {
+      launchLanguageServer(webSocket);
+    });
+  }
+});
