+++ Composer/packages/tools/language-servers/language-generation/src/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from './attachWSToHTTPServer';
+export * from './LGServer';
